Move RequestDecision folder
