+++ src/views/App/index.tsx
@@ -26,8 +26,8 @@ import GovernanceReviewTeam from 'views/GovernanceReviewTeam';
 import GovernanceTaskList from 'views/GovernanceTaskList';
 import GovernanceFeedback from 'views/GovernanceTaskList/Feedback';
 import LcidInfo from 'views/GovernanceTaskList/LcidInfo';
+import RequestDecision from 'views/GovernanceTaskList/RequestDecision';
 import GovernanceTaskListV1 from 'views/GovernanceTaskListV1';
-import RequestDecision from 'views/GovernanceTaskListV1/RequestDecision';
 import Help from 'views/Help';
 import Home from 'views/Home';
 import Login from 'views/Login';
